Added Resource Dao

+++ bring-a-smile-dal/src/main/java/com/bring/a/smile/model/FieldType.java
@@ -1,7 +1,7 @@
 package com.bring.a.smile.model;
 
 public enum FieldType {
-    INTEGER,
+    LONG,
     DATE,
     STRING,
